backend/requirements.txt
appdirs==1.4.4
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
CacheControl==0.13.1
cached-property==1.5.2
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
cryptography==3.4.8
decorator==4.4.2
defusedxml==0.7.1
Django==4.1.7
django-browser-reload==1.7.0
django-cors-headers==4.3.1
django-environ==0.9.0
django-redis==5.2.0
django-rest-knox==4.2.0
django-tailwind==3.5.0
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.16
ebaysdk==2.1.5
firebase-admin==6.3.0
freezegun==0.3.15
gevent==21.8.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-cloud-core==2.4.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
greenlet==1.1.2
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
httplib2==0.22.0
idna==2.10
isodate==0.6.1
Jinja2==2.11.3
jinja2-time==0.2.0
libsass==0.20.1
lxml==4.6.5
MarkupSafe==1.1.1
msgpack==1.0.7
num2words==0.5.9
ofxparse==0.21
passlib==1.7.4
Pillow==9.0.1
polib==1.1.0
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.8.0
psycopg2==2.9.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.8.0
pyodbc==4.0.39
pyOpenSSL==20.0.1
pyparsing==3.1.1
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.5
python-dateutil==2.8.1
python-slugify==8.0.1
python-stdnum==1.16
pytz==2023.3.post1
pyusb==1.0.2
pywin32==306
PyYAML==6.0
qrcode==6.1
redis==4.5.1
reportlab==3.5.59
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
tabulate==0.9.0
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.5
vobject==0.9.6.1
Werkzeug==2.0.2
WooCommerce==3.0.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==4.0.0
zope.event==4.6
zope.interface==5.5.2
